I think I figured out how to generate difference maps for cryoem inputs, still need to adjust functions to return double accum values and adjust functions responsible for generating coot scripts

+++ src/privateer/privateer-lib.cpp
@@ -817,7 +817,7 @@ bool privateer::util::read_coordinate_file_mtz (clipper::MMDBfile& mfile, clippe
 
 }
 
-bool privateer::util::read_coordinate_file_mrc (clipper::MMDBfile& mfile, clipper::MiniMol& mmol, clipper::String& ippdb, clipper::Xmap<float>& input_map, bool batch)
+bool privateer::util::read_coordinate_file_mrc (clipper::MMDBfile& mfile, clipper::MiniMol& mmol, clipper::String& ippdb, clipper::Xmap<double>& input_map, bool batch)
 {
     if (!batch)
     {
